requirements.txt
Flask>=3.0
Flask-Babel>=4.0
PyPDF2>=3.0
